backend/requirements.txt
fastapi
uvicorn[standard]

python-dotenv

requests
httpx

beautifulsoup4
html5lib

playwright

